src/main/java/com/leetcode/demo/easy/backtrack/PalindromePartitioning.java
package com.leetcode.demo.easy.backtrack;

import java.util.ArrayList;
import java.util.List;

/**
 * 给你一个字符串 s，请你将 s 分割成一些子串，使每个子串都是 回文串 。返回 s 所有可能的分割方案。
 * 回文串 是正着读和反着读都一样的字符串。
 * <p>
 * 来源：力扣（LeetCode）
 * 链接：https://leetcode-cn.com/problems/palindrome-partitioning
 * 著作权归领扣网络所有。商业转载请联系官方授权，非商业转载请注明出处。
 */
public class PalindromePartitioning {
    public List<List<String>> solve01(String s) {
        List<List<String>> res = new ArrayList<>();
        backtrack(res, new ArrayList<>(), s, 0);
        return res;
    }

    private void backtrack(List<List<String>> list, List<String> res, String s, int idx) {
        if (idx == s.length()) {
            list.add(new ArrayList<>(res));
        }

        for (int i = idx; i < s.length(); i++) {
            if (check(s, idx, i)) {
                res.add(s.substring(idx, i + 1));
                backtrack(list, res, s, i + 1);
                res.remove(res.size() - 1);
            }
        }
    }

    private boolean check(String s, int left, int right) {
        while (left < right) {
            if (s.charAt(left++) != s.charAt(right--)) {
                return false;
            }
        }
        return true;
    }
}
